% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleFun.R
\name{circleFun}
\alias{circleFun}
\title{Generate Perimeter Coordinates for Circle}
\usage{
circleFun(center = c(0, 0), diameter = 1, npoints = 100)
}
\arguments{
\item{center}{Where the center of the circle is defined}

\item{diameter}{Diameter of the circle}

\item{npoints}{Number of points to generate}
}
\description{
This function allows you to generate a dataframe with x and y coordinates that define the perimeter of a circle
}
\examples{
# 5 points that define a circle centered at (1,1) with a diameter of 2
circleFun(center = c(1,1), diameter = 2, npoints = 5)
}
\keyword{circle}
