% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derp.R
\name{derp}
\alias{derp}
\title{SELECT * From MySQL Table}
\usage{
derp(conn, x)
}
\arguments{
\item{conn}{Database connection}

\item{x}{Table name}
}
\description{
This function allows you to quickly run a SELECT * query.
}
\keyword{derp}
