% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Find the Mode of a Vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{Vector for which you want to calculate the mode}
}
\description{
This function allows you to quickly find the Mode of a vector
}
\examples{
v <- c(1,2,3,3,3,3,4,5)
getmode(v)
}
\keyword{to_under}
