% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag.R
\name{mag}
\alias{mag}
\title{Calculate Magnitude of Vector}
\usage{
mag(x)
}
\description{
This function allows you to quickly calculate the magnitude of a vector.
}
\examples{
x <- c(3,5,6)
mag(x)
}
\keyword{magnitude}
