% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mavback.R
\name{mavback}
\alias{mavback}
\title{Backwards moving average}
\usage{
mavback(x, n)
}
\arguments{
\item{x}{vector to compute rolling moving average}
}
\description{
This function allows you to compute the moving average of last N observations
}
\examples{
x <- 1:10
mavback(x)
}
\keyword{mavback}
