% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn.R
\name{mn}
\alias{mn}
\title{Convert Variable to Number}
\usage{
mn(x)
}
\arguments{
\item{x}{variable to convert to number}
}
\description{
This function allows you to quickly convert both strings and factors to numeric
}
\examples{
x <- "5"
mn(x)
}
\keyword{mn}
