% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cap.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalize First Letter of Each Word in String}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{String to capiltalize}
}
\description{
This function allows you to quickly capitalize the first letter of each word in a string.
}
\examples{
x <- "hello world"
simple_cap(x)

}
\keyword{simple_cap}
