% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_under.R
\name{to_under}
\alias{to_under}
\title{Convert Strings to Snake Case}
\usage{
to_under(x)
}
\arguments{
\item{x}{variable to convert to snake_case}
}
\description{
This function allows you to quickly convert strings to snake case
}
\examples{
x <- "Hello World"
to_under(x)
}
\keyword{to_under}
